Revert "Refactor : move testutils from `pkg` to `exoscale` to avoid import cycles"
This reverts commit 30d9747086306ce251f40e0247c29cae455639e5.
