+++ pkg/resources/iam/datasource_org_policy_test.go
@@ -3,7 +3,7 @@ package iam_test
 import (
 	"testing"
 
-	"github.com/exoscale/terraform-provider-exoscale/exoscale/testutils"
+	"github.com/exoscale/terraform-provider-exoscale/pkg/testutils"
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 )
