+++ pkg/testutils/testdata/resource_iam_api_key.tmpl
