+++ pkg/testutils/testdata/resource_iam_role.tmpl
