+++ pkg/testutils/testdata/resource_instance_pool.tmpl
