+++ pkg/testutils/testdata/resource_nlb.tmpl
