+++ pkg/testutils/testdata/resource_nlb_service.tmpl
